+++ src/main/java/ua/yakubovskiy/task3/entity/Shape.java
@@ -2,6 +2,6 @@ package ua.yakubovskiy.task3.entity;
 
 public interface Shape {
 
-    int getVolume();
+    double getVolume();
 
 }
